Refactor import paths for RecordingPreparedStatement; move to the correct package location
